Moved central init directory location
