+++ addons/ubat/src/core/mod.rs
@@ -2,8 +2,6 @@ pub mod game_manager;
 pub mod event_bus;
 pub mod config_manager;
 pub mod world_manager;
-pub mod initialization;
 
-pub use initialization::system_initializer;
 pub use event_bus::EventBus;
 pub use game_manager::GameManager;
