+++ addons/ubat/src/initialization/mod.rs
@@ -1,5 +1,7 @@
 pub mod system_initializer;
+pub mod world_initializer;
 pub mod configuration_service;
 
+pub use world_initializer::WorldInitializer;
 pub use configuration_service::ConfigurationService;
 pub use system_initializer::SystemInitializer;
